requirements.txt
pipecat-ai[silero,anthropic,deepgram,elevenlabs,groq,local]>=0.0.105
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
# Gmail API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
